requirements.txt
firebase-admin
RPi.GPIO
